+++ CHANGELOG.md
@@ -33,6 +33,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `opentelemetry-instrumentation-pymongo` `aggregate` and `getMore` capture statements support
   ([#3601](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3601))
 
+### Breaking changes
+
+- `opentelemetry-instrumentation-pymongo` introduce `db.operation`, refactor `db.statement`, refactor span name
+  ([#3606](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3606))
+
 ## Version 1.34.0/0.55b0 (2025-06-04)
 
 ### Fixed